+++ R/logic.R
@@ -293,8 +293,8 @@ select_next_item <- function(item_bank, opt) {
 }
 
 administer_next_item <- function(item_bank, show_item) {
-  psychTestR::reactive_page(
-    function(state, ...) {
+  list(
+    psychTestR::code_block(fun = function(state, ...) {
       test_state <- psychTestR::get_local("test_state", state)
       item_id <- test_state$next_item$item
       stopifnot(is.scalar.numeric(item_id),
@@ -302,7 +302,15 @@ administer_next_item <- function(item_bank, show_item) {
       item <- new_item(df = item_bank[item_id, ],
                        item_number = get_num_items_administered(test_state) + 1L,
                        num_items_in_test = get_num_items_in_test(test_state))
-      show_item(item, state, ...)
+      psychTestR::set_local("item", item, state)
+    }),
+    if (is.function(show_item)) {
+      psychTestR::reactive_page(function(state, ...) {
+        item <- psychTestR::get_local("item", state)
+        show_item(item, state, ...)
+      })
+    } else {
+      show_item
     }
   )
 }