+++ NEWS.md
@@ -1,3 +1,6 @@
+* `show_item` can now be a list of test elements rather than a function
+for a single reactive page
+
 # psychTestRCAT 1.1.0
 
 * Introducing compile_trial_by_trial_results()